+++ clang/lib/Sema/SemaChecking.cpp
@@ -3044,6 +3044,98 @@ static QualType getVectorElementType(ASTContext &Context, QualType VecTy) {
   return QualType();
 }
 
+static bool BuiltinElementwiseConvertFromArbitraryFP(Sema &S, CallExpr *TheCall,
+                                                     unsigned BuiltinID) {
+  if (S.checkArgCount(TheCall, 1))
+    return true;
+
+  std::string BuiltinName = S.Context.BuiltinInfo.getName(BuiltinID);
+  StringRef Suffix = BuiltinName;
+  if (!Suffix.consume_front("__builtin_elementwise_convert_from_"))
+    llvm_unreachable("unexpected builtin");
+
+  auto [SrcSuffix, DstSuffix] = Suffix.rsplit('_');
+  StringRef FormatName = llvm::StringSwitch<StringRef>(SrcSuffix)
+                             .Case("f8e5m2", "Float8E5M2")
+                             .Case("f8e4m3fn", "Float8E4M3FN")
+                             .Case("f6e3m2fn", "Float6E3M2FN")
+                             .Case("f6e2m3fn", "Float6E2M3FN")
+                             .Case("f4e2m1fn", "Float4E2M1FN")
+                             .Default("");
+  if (FormatName.empty())
+    llvm_unreachable("unknown source format");
+
+  QualType DstEltTy;
+  if (DstSuffix == "f16")
+    DstEltTy = S.Context.Float16Ty;
+  else if (DstSuffix == "bf16")
+    DstEltTy = S.Context.BFloat16Ty;
+  else if (DstSuffix == "f32")
+    DstEltTy = S.Context.FloatTy;
+  else if (DstSuffix == "f64")
+    DstEltTy = S.Context.DoubleTy;
+  else
+    llvm_unreachable("unknown destination type");
+
+  if (S.checkFloatingPointTypeSupport(DstEltTy, TheCall->getBeginLoc(),
+                                      /*DiagnoseTarget=*/true))
+    return true;
+
+  ExprResult ConvertedSrc = S.DefaultLvalueConversion(TheCall->getArg(0));
+  if (ConvertedSrc.isInvalid())
+    return true;
+  TheCall->setArg(0, ConvertedSrc.get());
+
+  QualType SrcTy = ConvertedSrc.get()->getType();
+  if (SrcTy->isDependentType()) {
+    TheCall->setType(S.Context.DependentTy);
+    return false;
+  }
+
+  if (SrcTy->isSizelessVectorType())
+    return S.Diag(TheCall->getBeginLoc(),
+                  diag::err_arbitrary_fp_sizeless_vector)
+           << "first" << BuiltinName;
+
+  QualType SrcEltTy = SrcTy;
+  const auto *SrcVecTy = SrcTy->getAs<VectorType>();
+  if (SrcVecTy)
+    SrcEltTy = SrcVecTy->getElementType();
+
+  if (SrcVecTy && !SrcTy->isExtVectorType() &&
+      SrcVecTy->getVectorKind() != VectorKind::Generic)
+    return S.Diag(TheCall->getBeginLoc(),
+                  diag::err_arbitrary_fp_unsupported_vector_type)
+           << "first" << BuiltinName << SrcTy;
+
+  if (!SrcEltTy->isIntegerType())
+    return S.Diag(TheCall->getBeginLoc(), diag::err_arbitrary_fp_non_int_type)
+           << "first" << BuiltinName;
+
+  if (SrcEltTy->isBooleanType() || SrcEltTy->isEnumeralType())
+    return S.Diag(TheCall->getBeginLoc(),
+                  diag::err_arbitrary_fp_invalid_int_type)
+           << "first" << BuiltinName;
+
+  unsigned FormatBits =
+      llvm::APFloatBase::getArbitraryFPFormatSizeInBits(FormatName);
+  if (S.Context.getIntWidth(SrcEltTy) != FormatBits)
+    return S.Diag(ConvertedSrc.get()->getBeginLoc(),
+                  diag::err_arbitrary_fp_int_width)
+           << (SrcVecTy != nullptr) << SrcEltTy << FormatBits << FormatName
+           << ConvertedSrc.get()->getSourceRange();
+
+  QualType DstTy = DstEltTy;
+  if (SrcVecTy)
+    DstTy =
+        SrcTy->isExtVectorType()
+            ? S.Context.getExtVectorType(DstEltTy, SrcVecTy->getNumElements())
+            : S.Context.getVectorType(DstEltTy, SrcVecTy->getNumElements(),
+                                      VectorKind::Generic);
+  TheCall->setType(DstTy);
+  return false;
+}
+
 ExprResult
 Sema::CheckBuiltinFunctionCall(FunctionDecl *FDecl, unsigned BuiltinID,
                                CallExpr *TheCall) {
@@ -3758,6 +3850,21 @@ Sema::CheckBuiltinFunctionCall(FunctionDecl *FDecl, unsigned BuiltinID,
   case Builtin::BI__builtin_elementwise_abs:
     if (PrepareBuiltinElementwiseMathOneArgCall(
             TheCall, EltwiseBuiltinArgTyRestriction::SignedIntOrFloatTy))
+      return ExprError();
+    break;
+
+#define ELEMENTWISE_CONVERT_FROM_CASES(Src)                                    \
+  case Builtin::BI__builtin_elementwise_convert_from_##Src##_f16:              \
+  case Builtin::BI__builtin_elementwise_convert_from_##Src##_bf16:             \
+  case Builtin::BI__builtin_elementwise_convert_from_##Src##_f32:              \
+  case Builtin::BI__builtin_elementwise_convert_from_##Src##_f64:
+    ELEMENTWISE_CONVERT_FROM_CASES(f8e5m2)
+    ELEMENTWISE_CONVERT_FROM_CASES(f8e4m3fn)
+    ELEMENTWISE_CONVERT_FROM_CASES(f6e3m2fn)
+    ELEMENTWISE_CONVERT_FROM_CASES(f6e2m3fn)
+    ELEMENTWISE_CONVERT_FROM_CASES(f4e2m1fn)
+#undef ELEMENTWISE_CONVERT_FROM_CASES
+    if (BuiltinElementwiseConvertFromArbitraryFP(*this, TheCall, BuiltinID))
       return ExprError();
     break;
 
@@ -6660,120 +6767,6 @@ ExprResult Sema::ConvertVectorExpr(Expr *E, TypeSourceInfo *TInfo,
 
   return ConvertVectorExpr::Create(Context, E, TInfo, DstTy, VK, OK, BuiltinLoc,
                                    RParenLoc, CurFPFeatureOverrides());
-}
-
-const StringLiteral *Sema::CheckArbitraryFPFormatArg(Expr *E) {
-  const auto *Literal = dyn_cast<StringLiteral>(E->IgnoreParenImpCasts());
-  if (!Literal) {
-    Diag(E->getBeginLoc(), diag::err_expr_not_string_literal)
-        << E->getSourceRange();
-    return nullptr;
-  }
-
-  if (!Literal->isOrdinary()) {
-    Diag(E->getBeginLoc(), diag::err_arbitrary_fp_non_ordinary_string_literal)
-        << E->getSourceRange();
-    return nullptr;
-  }
-
-  if (!llvm::APFloatBase::isValidArbitraryFPFormat(Literal->getString())) {
-    Diag(E->getBeginLoc(), diag::err_arbitrary_fp_invalid_format)
-        << Literal->getString() << E->getSourceRange();
-    return nullptr;
-  }
-
-  return Literal;
-}
-
-ExprResult Sema::ConvertFromArbitraryFPExpr(Expr *E, Expr *Format,
-                                            TypeSourceInfo *TInfo,
-                                            SourceLocation BuiltinLoc,
-                                            SourceLocation RParenLoc) {
-  const char *BuiltinName = "__builtin_convert_from_arbitrary_fp";
-
-  // The format is never dependent, so it can always be validated here.
-  const StringLiteral *FormatLiteral = CheckArbitraryFPFormatArg(Format);
-  if (!FormatLiteral)
-    return ExprError();
-  StringRef FormatName = FormatLiteral->getString();
-
-  ExprResult ConvertedSrc = DefaultLvalueConversion(E);
-  if (ConvertedSrc.isInvalid())
-    return ExprError();
-  E = ConvertedSrc.get();
-
-  QualType DstTy = TInfo->getType();
-  QualType SrcTy = E->getType();
-
-  if (!SrcTy->isDependentType()) {
-    if (SrcTy->isSizelessVectorType())
-      return ExprError(Diag(BuiltinLoc, diag::err_arbitrary_fp_sizeless_vector)
-                       << "first" << BuiltinName);
-
-    QualType SrcEltTy = SrcTy;
-    const auto *SrcVecTy = SrcTy->getAs<VectorType>();
-    if (SrcVecTy)
-      SrcEltTy = SrcVecTy->getElementType();
-
-    if (!SrcEltTy->isIntegerType())
-      return ExprError(Diag(BuiltinLoc, diag::err_arbitrary_fp_non_int_type)
-                       << "first" << BuiltinName);
-
-    unsigned FormatBits =
-        llvm::APFloatBase::getArbitraryFPFormatSizeInBits(FormatName);
-    if (Context.getIntWidth(SrcEltTy) != FormatBits)
-      return ExprError(Diag(E->getBeginLoc(), diag::err_arbitrary_fp_int_width)
-                       << (SrcVecTy != nullptr) << SrcEltTy << FormatBits
-                       << FormatName << E->getSourceRange());
-  }
-
-  if (!DstTy->isDependentType()) {
-    if (DstTy->isSizelessVectorType())
-      return ExprError(Diag(BuiltinLoc, diag::err_arbitrary_fp_sizeless_vector)
-                       << "third" << BuiltinName);
-
-    QualType DstEltTy = DstTy;
-    if (const auto *DstVecTy = DstTy->getAs<VectorType>())
-      DstEltTy = DstVecTy->getElementType();
-
-    if (!DstEltTy->isRealFloatingType() && !DstEltTy->isMFloat8Type())
-      return ExprError(Diag(BuiltinLoc, diag::err_arbitrary_fp_non_fp_type)
-                       << "third" << BuiltinName);
-
-    bool IsSupportedDstTy = !DstEltTy->isMFloat8Type();
-    if (IsSupportedDstTy) {
-      const llvm::fltSemantics &DstSem =
-          Context.getFloatTypeSemantics(DstEltTy);
-      IsSupportedDstTy = &DstSem == &llvm::APFloat::IEEEhalf() ||
-                         &DstSem == &llvm::APFloat::BFloat() ||
-                         &DstSem == &llvm::APFloat::IEEEsingle() ||
-                         &DstSem == &llvm::APFloat::IEEEdouble();
-    }
-
-    if (!IsSupportedDstTy)
-      return ExprError(
-          Diag(BuiltinLoc, diag::err_arbitrary_fp_unsupported_dst_type)
-          << DstTy << BuiltinName);
-  }
-
-  if (!SrcTy->isDependentType() && !DstTy->isDependentType()) {
-    const auto *SrcVecTy = SrcTy->getAs<VectorType>();
-    const auto *DstVecTy = DstTy->getAs<VectorType>();
-    if (DstVecTy && !SrcVecTy)
-      return ExprError(Diag(BuiltinLoc, diag::err_builtin_non_vector_type)
-                       << "first" << BuiltinName);
-    if (SrcVecTy && !DstVecTy)
-      return ExprError(Diag(BuiltinLoc, diag::err_builtin_non_vector_type)
-                       << "third" << BuiltinName);
-    if (SrcVecTy && DstVecTy &&
-        SrcVecTy->getNumElements() != DstVecTy->getNumElements())
-      return ExprError(
-          Diag(BuiltinLoc, diag::err_arbitrary_fp_incompatible_vector)
-          << BuiltinName);
-  }
-
-  return new (Context) clang::ConvertFromArbitraryFPExpr(
-      E, Format, TInfo, DstTy, VK_PRValue, OK_Ordinary, BuiltinLoc, RParenLoc);
 }
 
 bool Sema::BuiltinPrefetch(CallExpr *TheCall) {